+++ src/Vertex.zig
@@ -1,6 +1,6 @@
 const dvui = @import("dvui.zig");
 
-pub const Index = usize;
+pub const Index = u32;
 // pub const Index = switch (@import("build_options").vertex_index) {
 //     .u16 => u16,
 //     .u32 => u32,
